+++ src/main/java/com/engineal/musiclights/display/effects/FadeEffect.java
@@ -31,7 +31,7 @@ public class FadeEffect implements Effect {
     }
 
     @Override
-    public Color apply(int x) {
+    public Color apply(double x) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 }
